EntityExtraction/requirements.txt
appdirs==1.4.3
Cython==0.29.16
distlib==0.3.0
filelock==3.0.12
importlib-metadata==1.6.0
pdf2image==1.12.1
Pillow==7.1.1
pytesseract==0.3.3
six==1.12.0
tesseract==0.1.3
virtualenv==20.0.16
zipp==3.1.0
